Added ge Island list and items

+++ src/main/java/com/sdca/api/repository/IslandRepository.java
@@ -3,5 +3,5 @@ package com.sdca.api.repository;
 import com.sdca.api.model.Island;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IslandRepository extends CrudRepository<Island, Integer> {
+public interface IslandRepository extends CrudRepository<Island, Long> {
 }
